Create ReportDayValidator

+++ source/DailyReport.BusinessLogic/Servicies/ReportDayService.cs
@@ -1,8 +1,10 @@
-﻿using DailyReport.BusinessLogic.Interfaces;
+﻿using DailyReport.BusinessLogic.ExceptionValidators;
+using DailyReport.BusinessLogic.Interfaces;
 using DailyReport.BusinessLogic.Mappers;
 using DailyReport.BusinessLogic.ModelsDTO;
 using DailyReport.DataAccess.Interfaces;
 using DailyReport.DataAccess.Models;
+using static DailyReport.BusinessLogic.Exceptions.ExceptionValidator;
 
 namespace DailyReport.BusinessLogic.Servicies
 {
@@ -19,8 +21,11 @@ namespace DailyReport.BusinessLogic.Servicies
         {
             if (item == null)
             {
-                //
-                throw new ArgumentNullException("item");
+                throw new ValidationException("Can not create a report day");
+            }
+            if (ReportDayValidator.ReportDayExists(item.RecordDay, item.UserName, _reportDayRepository))
+            {
+                throw new ValidationException($"Day '{item.RecordDay}' already exists");
             }
             ReportDay reportDay = ReportDayMapper.FromDTO(item);
             await _reportDayRepository.CreateAsync(reportDay);
@@ -28,6 +33,14 @@ namespace DailyReport.BusinessLogic.Servicies
 
         public async Task DeleteAsync(ReportDayDTO item)
         {
+            if (item == null)
+            {
+                throw new ValidationException("Can not delete a day");
+            }
+            if (item.Id <= 0)
+            {
+                throw new ValidationException("It is impossible");
+            }
             ReportDay reportDay = ReportDayMapper.FromDTO(item);
             await _reportDayRepository.DeleteAsync(reportDay);
         }
@@ -46,6 +59,10 @@ namespace DailyReport.BusinessLogic.Servicies
 
         public async Task<ReportDayDTO> GetByIdAsync(int? id)
         {
+            if (id <= 0)
+            {
+                throw new ValidationException("It is impossible");
+            }
             var reportDay = await _reportDayRepository.GetByIdAsync(id);
             ReportDayDTO reportDayDTO = ReportDayMapper.ToDTO(reportDay);
             return reportDayDTO;
@@ -53,6 +70,14 @@ namespace DailyReport.BusinessLogic.Servicies
 
         public async Task UpdateAsync(ReportDayDTO item)
         {
+            if (item == null)
+            {
+                throw new ValidationException("Can not update a day");
+            }
+            if (item.Id <= 0)
+            {
+                throw new ValidationException("It is impossible");
+            }
             ReportDay reportDay = ReportDayMapper.FromDTO(item);
             await _reportDayRepository.UpdateAsync(reportDay);
         }
